requirements.txt
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
sentencepiece
numpy>=1.21.0
opencv-python>=4.5.0
Pillow>=8.3.0
tqdm>=4.62.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
streamlit>=1.28.0
fpdf2>=2.7.0
requests>=2.25.0
pandas>=1.3.0
